Ensure hash always returns a Fixnum. Previous implementation could return Bignum.
The docs for Object#hash seem to suggest that it's intended that
hash returns a Fixnum, though is less clear about whether it's
a hard requirement. It suggests the Hash class will work around
a longer hash value, but I'm unsure whether other classes (if there
are any that care about hash) would do the same. Also, I'm not
sure how appropriate truncation is as a workaround, but then
maybe Array#hash just does truncation anyway.